+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added tools for loading models from a given description struct `RobotSpec` (https://github.com/Simple-Robotics/candlewick/pull/37)
 - Added Candlewick visualizer runtime, to be used to (up)load Pinocchio models and submit states to be displayed asynchronously. (https://github.com/Simple-Robotics/candlewick/pull/37)
 - Embed Inter Medium font into the application (https://github.com/Simple-Robotics/candlewick/pull/37)
 
@@ -21,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `core/Device.h` : fix release() function
 - Python examples : wrap some import statements in try/except statements (https://github.com/Simple-Robotics/candlewick/pull/37)
+
+### Removed
+
+- Remove dependency on `robot_descriptions_cpp` (https://github.com/Simple-Robotics/candlewick/pull/37)
 
 ## [0.7.0] - 2025-07-04